skeleton_backend/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
joblib==0.13.2
numpy==1.17.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.0
sklearn==0.0
sqlparse==0.3.0
django-extensions==2.2.1
django-rest-knox==4.1.0
drf-yasg==1.16.1
packaging==19.2
flex==6.14.1
dj-database-url==0.5.0
whitenoise==3.3.1
psycopg2==2.6.2